Added harvest buttons to the tile menu
If the selected tile has a tree or shrub on it, a harvest button for
that resource will appear in the tile menu. Adjusted the anchor point of
trees. Made sprites in the 'layered' group render in order of y-value.
Added icons for the different ways of harvesting (chopping, harvesting).